projectbeginnetje
projectstructuur met gradle als bouwsysteem
