--- CGenerator/src/main/java/org/example/IExpressionGenerator.java
@@ -1,5 +1,0 @@
-package org.example;
-
-public interface IExpressionGenerator {
-    String getNewExpression(int currentDepth, DataType type, boolean terminating);
-}
