--- CGenerator/src/main/java/org/example/IFunctionGenerator.java
@@ -1,7 +1,0 @@
-package org.example;
-
-import java.util.List;
-
-public interface IFunctionGenerator {
-    Function getNewFunction(DataType type);
-}
