--- CGenerator/src/main/java/org/example/IGlobalVariableGenerator.java
@@ -1,5 +1,0 @@
-package org.example;
-
-public interface IGlobalVariableGenerator {
-    Variable getNewGlobalVariable(DataType type);
-}
