--- CGenerator/src/main/java/org/example/IStatementGenerator.java
@@ -1,5 +1,0 @@
-package org.example;
-
-public interface IStatementGenerator {
-    String getNewStatement();
-}
