--- CGenerator/src/main/java/org/example/IStructGenerator.java
@@ -1,5 +1,0 @@
-package org.example;
-
-public interface IStructGenerator {
-    Struct getNewStruct();
-}
