+++ build.ps1
@@ -1,0 +1,2 @@
+gradle producer_jar
+gradle assessor_jar
