+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Sun Sep 03 18:49:31 CEST 2023
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.2-bin.zip
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
