+++ run_producer.ps1
@@ -1,0 +1,2 @@
+# Het idee hiervan is dat de map waar de containers moeten komen te staan als argument wordt meegegeven.
+java -jar build/libs/producer.jar "containers"
