+++ src/main/java/nl/ou/debm/assessor/Main.java
@@ -1,0 +1,15 @@
+package nl.ou.debm.assessor;
+
+import nl.ou.debm.common.*;
+
+// Press Shift twice to open the Search Everywhere dialog and type `show whitespaces`,
+// then press Enter. You can now see whitespace characters in your code.
+public class Main {
+    public static void main(String[] args) {
+        // Press Alt+Enter with your caret at the highlighted text to see how
+        // IntelliJ IDEA suggests fixing it.
+        System.out.println("Dit is de assessor");
+        System.out.println("Gebruik van common: ");
+        System.out.println(lib.inc(1));
+    }
+}
