+++ src/main/java/nl/ou/debm/common/lib.java
@@ -1,0 +1,7 @@
+package nl.ou.debm.common;
+
+public class lib {
+    public static int inc(int i) {
+        return i+1;
+    }
+}
