+++ src/main/java/nl/ou/debm/producer/IExpressionGenerator.java
@@ -1,0 +1,5 @@
+package nl.ou.debm.producer;
+
+public interface IExpressionGenerator {
+    String getNewExpression(int currentDepth, DataType type, boolean terminating);
+}
