+++ src/main/java/nl/ou/debm/producer/IFunctionGenerator.java
@@ -1,0 +1,7 @@
+package nl.ou.debm.producer;
+
+import java.util.List;
+
+public interface IFunctionGenerator {
+    Function getNewFunction(DataType type);
+}
