+++ src/main/java/nl/ou/debm/producer/IGlobalVariableGenerator.java
@@ -1,0 +1,5 @@
+package nl.ou.debm.producer;
+
+public interface IGlobalVariableGenerator {
+    Variable getNewGlobalVariable(DataType type);
+}
