+++ src/main/java/nl/ou/debm/producer/IStatementGenerator.java
@@ -1,0 +1,5 @@
+package nl.ou.debm.producer;
+
+public interface IStatementGenerator {
+    String getNewStatement();
+}
