+++ src/main/java/nl/ou/debm/producer/IStructGenerator.java
@@ -1,0 +1,5 @@
+package nl.ou.debm.producer;
+
+public interface IStructGenerator {
+    Struct getNewStruct();
+}
